requirements-macos-metal-intel.txt
tensorflow-macos==2.8.0
tensorflow-metal==0.3
pyside6==6.2.2
pillow==8.4
